refactor: getLikeKeyPrefix 메서드 옮기기
